pg_locale.c, unicode_case.c: use size_t for iteration.
Already done by e615da8cb21b in master. Backpatch these particular
cases to make other backpatches in this area safer, and to be safe for
callers in extensions that we don't know about.

This applies to REL_19_STABLE and REL_18_STABLE only.

Suggested-by: Andres Freund <andres@anarazel.de>
Discussion: https://postgr.es/m/v36ssaygf7grb3qzfsjhtdzi7kqd45ds56nyuf7gi5qjml4qbb@ezmfqzmhlrs2

+++ src/common/unicode_case.c
@@ -341,7 +341,7 @@ check_final_sigma(const unsigned char *str, size_t len, size_t offset)
 	int			ulen;
 
 	/* iterate backwards looking for preceding character */
-	for (int i = offset; i > 0;)
+	for (size_t i = offset; i > 0;)
 	{
 		/* skip backwards through continuation bytes */
 		i--;
@@ -369,7 +369,7 @@ check_final_sigma(const unsigned char *str, size_t len, size_t offset)
 	ulen = utf8_mblen((const unsigned char *) str + offset);
 
 	/* iterate forward looking for following character */
-	for (int i = offset + ulen; i < len;)
+	for (size_t i = offset + ulen; i < len;)
 	{
 		ulen = utf8_mblen((const unsigned char *) str + i);
